+++ src/main/java/org/example/model/Hero.java
@@ -4,26 +4,26 @@ import lombok.Data;
 
 @Data
 public class Hero {
-    private int baseSpeed;
-    private int basePower;
-    private int baseRange;
-    private int levelSpeedCoeff;
-    private int levelPowerCoeff;
-    private int levelRangeCoeff;
-    private int x;
-    private int y;
-    private int level;
-    private int exp;
+    private long baseSpeed;
+    private long basePower;
+    private long baseRange;
+    private long levelSpeedCoeff;
+    private long levelPowerCoeff;
+    private long levelRangeCoeff;
+    private long x;
+    private long y;
+    private long level;
+    private long exp;
 
-    public int getSpeed() {
-        return (int) Math.floor((double) baseSpeed * (1 + (double) level * levelSpeedCoeff / 100.0));
+    public long getSpeed() {
+        return (long) Math.floor((double) baseSpeed * (1 + (double) level * levelSpeedCoeff / 100.0));
     }
 
-    public int getPower() {
-        return (int) Math.floor((double) basePower * (1 + (double) level * levelPowerCoeff / 100.0));
+    public long getPower() {
+        return (long) Math.floor((double) basePower * (1 + (double) level * levelPowerCoeff / 100.0));
     }
 
-    public int getRange() {
-        return (int) Math.floor((double) baseRange * (1 + (double) level * levelRangeCoeff / 100.0));
+    public long getRange() {
+        return (long) Math.floor((double) baseRange * (1 + (double) level * levelRangeCoeff / 100.0));
     }
 }
